requirements.txt
python
pandas
python-dotenv
requests
geopandas